clang-tidy sweep: bounded string ops, narrowing casts, header self-containment
GitOTA / WResp: replace strcat/strcpy with strlcat/strlcpy (and sprintf
with snprintf) where a destination size is in scope. WResp's escapeString
now takes an explicit escapedSize parameter and bounds-checks the
default-case byte write; all three callers pass the available remaining
space. Also fix a real off-by-one in JsonFormatter::calcEscapedLength —
the original loop iterated raw[strlen(raw)] down to raw[1], reading the
null terminator and skipping raw[0]. Silently under-counted by 1 when
the value started with an escapable char, which let the precondition
in _safecat under-allocate.

setFirmwareFile() in GitOTA now covers s2/s3/c2/c3/c6/h2 (matching the
chip set advertised by setAssetProperty) and initialises the chosen
binary name inline so unknown future chip enums fall through to the
safe esp32.bin default rather than reading uninitialised memory.

MQTT::callback: explicitly cast getMyPos() to int8_t at the
setMyPosition call site; the implicit float→int8_t conversion was
tripping bugprone-narrowing-conversions.

ShadeConfigFile / Utils: floor() of a float promoted to double — switch
to floorf(). In Utils::formatISO, drop the floor() entirely since the
argument was integer-divided int (floor of an int is a no-op). In
calcTZOffset, cast mktime()-mktime() to long explicitly; the offset
fits in 32 bits.

Header self-containment: ControllerNetwork.h now includes <WiFi.h> and
ConfigSettings.h directly instead of relying on transitive includes.
app_main.h is deleted: it forward-declared a single static helper used
only inside app_main.cpp; declaration moved into the .cpp and the
definition marked static to match.

clang-tidy config: add bugprone-signed-char-misuse.CharTypdefsToIgnore
= int8_t (note the misspelling — clang-tidy 19.x has a typo in the
option name; revisit when the toolchain advances). Disable
clang-diagnostic-unused-command-line-argument so the GCC-only
-mlongcalls flag stops generating noise. Move CheckOptions above Checks
and rewrite Checks as a plain quoted string (the folded-scalar form was
swallowing CheckOptions on this clang-tidy version).

+++ main/MQTT.cpp
@@ -137,7 +137,7 @@ void MQTTClass::receive(const char *topic, byte*payload, uint32_t length) {
       }
       else if(strncmp(command, "myTiltPos", sizeof(command)) == 0) {
         if(val >= 0 && val <= 100)
-          shade->setMyPosition(shade->getMyPos(), static_cast<int8_t>(val));
+          shade->setMyPosition(static_cast<int8_t>(shade->getMyPos()), static_cast<int8_t>(val));
       }
       else if(strncmp(command, "sunFlag", sizeof(command)) == 0) {
         if(val > 0) somfy.enqueueShadeCommand(shadeId, somfy_commands::SunFlag, shade->repeats);
